requirements.txt
flask
bcrypt
pyjwt
python-dotenv
